@@ -12,15 +12,15 @@
         "email": "dmitry.g.ivanov@gmail.com"
     }],
     "require": {
-        "php": "^8.1",
-        "guzzlehttp/guzzle": "^7.5",
-        "guzzlehttp/psr7": "^2.0",
-        "illuminate/support": "^10.0"
+        "php": "^8.2",
+        "guzzlehttp/guzzle": "^7.8.2",
+        "guzzlehttp/psr7": "^2.4",
+        "illuminate/support": "^11.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^10.5",
-        "mockery/mockery": "^1.5.1",
-        "orchestra/testbench": "^8.0"
+        "phpunit/phpunit": "^11.3.6",
+        "mockery/mockery": "^1.6.10",
+        "orchestra/testbench": "^9.11.2"
     },
     "autoload": {
         "psr-4": {
