@@ -7,29 +7,30 @@ use Illuminated\Wikipedia\Grabber\Formatter\BulmaFormatter;
 use Illuminated\Wikipedia\Grabber\Formatter\Formatter;
 use Illuminated\Wikipedia\Grabber\Formatter\PlainFormatter;
 use Illuminated\Wikipedia\Tests\TestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 class FormatterTest extends TestCase
 {
-    /** @test */
-    public function it_has_static_factory_method_which_returns_plain_formatter_by_default()
+    #[Test]
+    public function it_has_static_factory_method_which_returns_plain_formatter_by_default(): void
     {
         $this->assertInstanceOf(PlainFormatter::class, Formatter::factory('foobar', collect()));
     }
 
-    /** @test */
-    public function and_it_returns_plain_formatter_if_asked()
+    #[Test]
+    public function and_it_returns_plain_formatter_if_asked(): void
     {
         $this->assertInstanceOf(PlainFormatter::class, Formatter::factory('plain', collect()));
     }
 
-    /** @test */
-    public function and_it_returns_bulma_formatter_if_asked()
+    #[Test]
+    public function and_it_returns_bulma_formatter_if_asked(): void
     {
         $this->assertInstanceOf(BulmaFormatter::class, Formatter::factory('bulma', collect()));
     }
 
-    /** @test */
-    public function and_it_returns_bootstrap_formatter_if_asked()
+    #[Test]
+    public function and_it_returns_bootstrap_formatter_if_asked(): void
     {
         $this->assertInstanceOf(BootstrapFormatter::class, Formatter::factory('bootstrap', collect()));
     }
