@@ -2,46 +2,37 @@
 
 namespace Illuminated\Wikipedia\Tests\Grabber;
 
-use Illuminated\Wikipedia\Grabber\Page;
-use Illuminated\Wikipedia\Grabber\Preview;
 use Illuminated\Wikipedia\Tests\TestCase;
 use Illuminated\Wikipedia\Wikipedia;
+use PHPUnit\Framework\Attributes\Test;
 
 class RandomTest extends TestCase
 {
-    /** @test */
-    public function it_can_grab_random_page_for_en()
+    #[Test]
+    public function it_can_grab_random_page_for_en(): void
     {
         $page = (new Wikipedia)->randomPage();
-
-        $this->assertInstanceOf(Page::class, $page);
         $this->assertTrue($page->isSuccess());
     }
 
-    /** @test */
-    public function it_can_grab_random_page_for_ru()
+    #[Test]
+    public function it_can_grab_random_page_for_ru(): void
     {
         $page = (new Wikipedia('ru'))->randomPage();
-
-        $this->assertInstanceOf(Page::class, $page);
         $this->assertTrue($page->isSuccess());
     }
 
-    /** @test */
-    public function it_can_grab_random_preview_for_en()
+    #[Test]
+    public function it_can_grab_random_preview_for_en(): void
     {
         $preview = (new Wikipedia)->randomPreview();
-
-        $this->assertInstanceOf(Preview::class, $preview);
         $this->assertTrue($preview->isSuccess());
     }
 
-    /** @test */
-    public function it_can_grab_random_preview_for_ru()
+    #[Test]
+    public function it_can_grab_random_preview_for_ru(): void
     {
         $preview = (new Wikipedia('ru'))->randomPreview();
-
-        $this->assertInstanceOf(Preview::class, $preview);
         $this->assertTrue($preview->isSuccess());
     }
 }
