@@ -4,11 +4,12 @@ namespace Illuminated\Wikipedia\Tests\Grabber\Wikitext\Normalizer;
 
 use Illuminated\Wikipedia\Grabber\Wikitext\Normalizer\LocaleFile;
 use Illuminated\Wikipedia\Tests\TestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 class LocaleFileTest extends TestCase
 {
-    /** @test */
-    public function it_has_normalize_method_which_handles_locale_files()
+    #[Test]
+    public function it_has_normalize_method_which_handles_locale_files(): void
     {
         $this->assertEquals(
             trim(file_get_contents(__DIR__ . '/LocaleFileTest/normalize.txt')),
