@@ -4,11 +4,12 @@ namespace Illuminated\Wikipedia\Tests\Grabber\Wikitext\Templates;
 
 use Illuminated\Wikipedia\Grabber\Wikitext\Templates\ListenTemplate;
 use Illuminated\Wikipedia\Tests\TestCase;
+use PHPUnit\Framework\Attributes\Test;
 
 class ListenTemplateTest extends TestCase
 {
-    /** @test */
-    public function it_has_transform_method_which_works_if_there_is_no_title_and_description()
+    #[Test]
+    public function it_has_transform_method_which_works_if_there_is_no_title_and_description(): void
     {
         $template = new ListenTemplate('{{Listen | type = music | filename = Accordion chords-01.ogg | footer = Some footer }}');
 
@@ -18,8 +19,8 @@ class ListenTemplateTest extends TestCase
         );
     }
 
-    /** @test */
-    public function and_it_works_for_the_case_when_only_title_set()
+    #[Test]
+    public function and_it_works_for_the_case_when_only_title_set(): void
     {
         $template = new ListenTemplate('{{Listen | type = music | filename = Accordion chords-01.ogg | Title = Accordion chords | footer = Some footer }}');
 
@@ -29,8 +30,8 @@ class ListenTemplateTest extends TestCase
         );
     }
 
-    /** @test */
-    public function and_it_works_for_the_case_when_only_description_set()
+    #[Test]
+    public function and_it_works_for_the_case_when_only_description_set(): void
     {
         $template = new ListenTemplate('{{Listen | type = music | filename = Accordion chords-01.ogg | Description = Chords being played on an accordion | footer = Some footer }}');
 
@@ -40,8 +41,8 @@ class ListenTemplateTest extends TestCase
         );
     }
 
-    /** @test */
-    public function and_it_works_for_the_case_when_title_and_description_are_set()
+    #[Test]
+    public function and_it_works_for_the_case_when_title_and_description_are_set(): void
     {
         $template = new ListenTemplate('{{Listen | type = music | filename = Accordion chords-01.ogg | Title = Accordion chords | Description = Chords being played on an accordion | footer = Some footer }}');
 
