@@ -4,19 +4,21 @@ namespace Illuminated\Wikipedia\Tests;
 
 use GuzzleHttp\Client;
 use Illuminated\Wikipedia\Wikipedia;
+use PHPUnit\Framework\Attributes\Test;
 
 class GrabberTest extends TestCase
 {
-    /** @test */
-    public function it_has_get_client_method()
+    #[Test]
+    public function it_has_get_client_method(): void
     {
         $client = (new Wikipedia)->getClient();
 
+        /** @noinspection UnnecessaryAssertionInspection */
         $this->assertInstanceOf(Client::class, $client);
     }
 
-    /** @test */
-    public function client_has_default_user_agent_logic()
+    #[Test]
+    public function client_has_default_user_agent_logic(): void
     {
         config([
             'wikipedia-grabber.user_agent' => null,
@@ -28,18 +30,18 @@ class GrabberTest extends TestCase
 
         $this->assertEquals(
             'Laravel Wikipedia Grabber (https://github.com/dmitry-ivanov/laravel-wikipedia-grabber)',
-            $client->getConfig('headers')['User-Agent']
+            $client->getConfig('headers')['User-Agent'],
         );
     }
 
-    /** @test */
-    public function and_it_takes_specified_user_agent_if_set()
+    #[Test]
+    public function and_it_takes_specified_user_agent_if_set(): void
     {
         $client = (new Wikipedia)->getClient();
 
         $this->assertEquals(
             'Laravel Wikipedia Grabber (https://github.com/dmitry-ivanov/laravel-wikipedia-grabber; dmitry.g.ivanov@gmail.com)',
-            $client->getConfig('headers')['User-Agent']
+            $client->getConfig('headers')['User-Agent'],
         );
     }
 }
